function-data-meeting/requirements.txt
functions-framework==3.*
datetime==5.*
requests==2.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
pandas==2.*